mod: 移动代码到单独文件里，保持代码干净
